Move maze creation to its own file
