+++ PixelMaze/PixelMaze.js
@@ -3,9 +3,9 @@ import { Grid, Index2D } from "../sketchlib/Grid.js";
 import { GridDirection } from "../sketchlib/GridDiection.js";
 import { generate_maze } from "../sketchlib/RandomDFSMaze.js";
 import { blit_sprite, blit_tilemap, P5Sprite, P5Tilemap } from "./blit.js";
+import { make_maze } from "./make_maze.js";
 import { parse_resources } from "./parse_resources.js";
 import { preload_p5_resources } from "./preload.js";
-import { Sprite } from "./Sprite.js";
 import { Tilemap } from "./Tilemap.js";
 
 const TILE_SIZE = 16;
@@ -13,109 +13,7 @@ const TILE_SCALE = 2;
 
 const MAZE_ROWS = 4;
 const MAZE_COLS = 3;
-const MAZE = generate_maze(MAZE_ROWS, MAZE_COLS);
-
-const Tiles = {
-  CEILING: 0,
-  WALL: 1,
-  FLOOR: 2,
-};
-
-const INDICES = new Grid(3 * (2 * MAZE_ROWS - 1), 3 * (2 * MAZE_COLS - 1));
-INDICES.fill(() => Tiles.CEILING);
-
-MAZE.for_each((index, cell) => {
-  const { i: cell_y, j: cell_x } = index;
-
-  // In the output grid, each input grid cell is 3 cells wide, + another 3
-  // for the corridor between each cell
-  const row_offset = 6 * cell_y;
-  const col_offset = 6 * cell_x;
-
-  // First, we need to add a 3x3 chunk of floor tiles for this cell
-  for (let i = 0; i < 3; i++) {
-    const dst_row = row_offset + i;
-    for (let j = 0; j < 3; j++) {
-      const dst_col = col_offset + j;
-      INDICES.set(new Index2D(dst_row, dst_col), Tiles.FLOOR);
-    }
-  }
-
-  // If we're connected on the right, draw a corridor. 3 floor tiles, and
-  // 6 wall tiles above it
-  if (cell.is_connected(GridDirection.RIGHT)) {
-    // Middle of the 3 tiles, hence + 1
-    const dst_row = row_offset + 1;
-    for (let i = 0; i < 3; i++) {
-      const dst_col = col_offset + 3 + i;
-
-      // Draw one floor tile
-      const floor_index = new Index2D(dst_row, dst_col);
-      INDICES.set(floor_index, Tiles.FLOOR);
-
-      // The two tiles above it are the wall. Since the corridor is in the
-      // middle, the bottom of the wall will always fit in the grid, but the
-      // top might be cropped a bit.
-      const wall_bottom = floor_index.up();
-      INDICES.set(wall_bottom, Tiles.WALL);
-
-      const wall_top = wall_bottom.up();
-      if (wall_top) {
-        INDICES.set(wall_top, Tiles.WALL);
-      }
-    }
-  }
-
-  // The vertical corridors are similar, except the walls are placed a bit
-  // differently
-  if (cell.is_connected(GridDirection.DOWN)) {
-    const dst_col = col_offset + 1;
-    for (let i = 0; i < 3; i++) {
-      const dst_row = row_offset + 3 + i;
-
-      const floor_index = new Index2D(dst_row, dst_col);
-      INDICES.set(floor_index, Tiles.FLOOR);
-
-      if (i > 0) {
-        const left_wall = floor_index.left();
-        INDICES.set(left_wall, Tiles.WALL);
-
-        const right_wall = floor_index.right();
-        INDICES.set(right_wall, Tiles.WALL);
-      }
-    }
-  }
-
-  // For cells below us, we still need a wall above the room even if there
-  // was no connection. Each wall is 2 tiles high, so we write a 2x3 rectangle
-  // of cells
-  if (cell_y + 1 < MAZE.rows && !cell.is_connected(GridDirection.DOWN)) {
-    for (let i = 1; i < 3; i++) {
-      const dst_row = row_offset + 3 + i;
-      for (let j = 0; j < 3; j++) {
-        const dst_col = col_offset + j;
-        INDICES.set(new Index2D(dst_row, dst_col), Tiles.WALL);
-      }
-    }
-  }
-});
-
-/**
- * Draw a tilemap on the screen
- * @param {any} p p5 context
- * @param {Tilemap} tilemap The tilemap
- * @param {Point} origin The pixel coords of where the top left corner of the grid will appear
- * @param {number} scale The scale factor for upscaling tiles
- */
-/*
-function blit_tilemap(p, tilemap, origin, scale) {
-  const { tileset, indices } = tilemap;
-  indices.for_each((index, tile_id) => {
-    const { i, j } = index;
-    const position = Point.direction(j, i).scale(scale * tileset.tile_size);
-    blit_tile(p, tileset, tile_id, origin.add(position), scale);
-  });
-}*/
+const INDICES = make_maze(MAZE_ROWS, MAZE_COLS);
 
 const ORIGIN_CHARACTER = Point.direction(0, TILE_SIZE);
 const RESOURCE_MANIFEST = {
